Don't attribute conntracked connections to k8s pause containers.
