+++ probe/docker/registry.go
@@ -1,6 +1,8 @@
 package docker
 
 import (
+	"fmt"
+	"strings"
 	"sync"
 	"time"
 
@@ -400,4 +402,15 @@ func (r *registry) WalkImages(f func(*docker_client.APIImages)) {
 		}
 		return false
 	})
+}
+
+// ImageNameWithoutVersion splits the image name apart, returning the name
+// without the version, if possible
+func ImageNameWithoutVersion(name string) string {
+	parts := strings.SplitN(name, "/", 3)
+	if len(parts) == 3 {
+		name = fmt.Sprintf("%s/%s", parts[1], parts[2])
+	}
+	parts = strings.SplitN(name, ":", 2)
+	return parts[0]
 }
