+++ probe/docker/registry_test.go
@@ -451,3 +451,18 @@ func TestRegistryDelete(t *testing.T) {
 		}
 	})
 }
+
+func TestDockerImageName(t *testing.T) {
+	for _, input := range []struct{ in, name string }{
+		{"foo/bar", "foo/bar"},
+		{"foo/bar:baz", "foo/bar"},
+		{"reg:123/foo/bar:baz", "foo/bar"},
+		{"docker-registry.domain.name:5000/repo/image1:ver", "repo/image1"},
+		{"foo", "foo"},
+	} {
+		name := docker.ImageNameWithoutVersion(input.in)
+		if name != input.name {
+			t.Fatalf("%s: %s != %s", input.in, name, input.name)
+		}
+	}
+}
